# openapi/photo_album/codegen_client/openapi_client/models/photo.py
# coding: utf-8

"""
    Sample OpenAPI Specification

    An OpenAPI specification sample for [Build, Deploy, and Manage Networked APIs: An Overview](https://goo.gl/VardtG) document.  # noqa: E501

    OpenAPI spec version: 0.0.1
    Generated by: https://openapi-generator.tech
"""


import pprint
import re  # noqa: F401

import six


class Photo(object):
    """NOTE: This class is auto generated by OpenAPI Generator.
    Ref: https://openapi-generator.tech

    Do not edit the class manually.
    """

    """
    Attributes:
      openapi_types (dict): The key is attribute name
                            and the value is attribute type.
      attribute_map (dict): The key is attribute name
                            and the value is json key in definition.
    """
    openapi_types = {
        'name': 'str',
        'display_name': 'str',
        'created_at': 'int',
        'data': 'str'
    }

    attribute_map = {
        'name': 'name',
        'display_name': 'display_name',
        'created_at': 'created_at',
        'data': 'data'
    }

    def __init__(self, name=None, display_name=None, created_at=None, data=None):  # noqa: E501
        """Photo - a model defined in OpenAPI"""  # noqa: E501

        self._name = None
        self._display_name = None
        self._created_at = None
        self._data = None
        self.discriminator = None

        if name is not None:
            self.name = name
        if display_name is not None:
            self.display_name = display_name
        if created_at is not None:
            self.created_at = created_at
        self.data = data

    @property
    def name(self):
        """Gets the name of this Photo.  # noqa: E501


        :return: The name of this Photo.  # noqa: E501
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name):
        """Sets the name of this Photo.


        :param name: The name of this Photo.  # noqa: E501
        :type: str
        """

        self._name = name

    @property
    def display_name(self):
        """Gets the display_name of this Photo.  # noqa: E501


        :return: The display_name of this Photo.  # noqa: E501
        :rtype: str
        """
        return self._display_name

    @display_name.setter
    def display_name(self, display_name):
        """Sets the display_name of this Photo.


        :param display_name: The display_name of this Photo.  # noqa: E501
        :type: str
        """

        self._display_name = display_name

    @property
    def created_at(self):
        """Gets the created_at of this Photo.  # noqa: E501


        :return: The created_at of this Photo.  # noqa: E501
        :rtype: int
        """
        return self._created_at

    @created_at.setter
    def created_at(self, created_at):
        """Sets the created_at of this Photo.


        :param created_at: The created_at of this Photo.  # noqa: E501
        :type: int
        """

        self._created_at = created_at

    @property
    def data(self):
        """Gets the data of this Photo.  # noqa: E501


        :return: The data of this Photo.  # noqa: E501
        :rtype: str
        """
        return self._data

    @data.setter
    def data(self, data):
        """Sets the data of this Photo.


        :param data: The data of this Photo.  # noqa: E501
        :type: str
        """
        if data is None:
            raise ValueError("Invalid value for `data`, must not be `None`")  # noqa: E501

        self._data = data

    def to_dict(self):
        """Returns the model properties as a dict"""
        result = {}

        for attr, _ in six.iteritems(self.openapi_types):
            value = getattr(self, attr)
            if isinstance(value, list):
                result[attr] = list(map(
                    lambda x: x.to_dict() if hasattr(x, "to_dict") else x,
                    value
                ))
            elif hasattr(value, "to_dict"):
                result[attr] = value.to_dict()
            elif isinstance(value, dict):
                result[attr] = dict(map(
                    lambda item: (item[0], item[1].to_dict())
                    if hasattr(item[1], "to_dict") else item,
                    value.items()
                ))
            else:
                result[attr] = value

        return result

    def to_str(self):
        """Returns the string representation of the model"""
        return pprint.pformat(self.to_dict())

    def __repr__(self):
        """For `print` and `pprint`"""
        return self.to_str()

    def __eq__(self, other):
        """Returns true if both objects are equal"""
        if not isinstance(other, Photo):
            return False

        return self.__dict__ == other.__dict__

    def __ne__(self, other):
        """Returns true if both objects are not equal"""
        return not self == other